+++ src/main/java/com/github/propheticeclipse/tensurastarlight/ability/skill/intrinsic/arsnoveau/wildenStalkerGeneticsSkill.java
@@ -2,17 +2,22 @@ package com.github.propheticeclipse.tensurastarlight.ability.skill.intrinsic.ars
 
 import com.github.propheticeclipse.tensurastarlight.config.skills.arsnouveauSeriesSkillConfig;
 import com.github.propheticeclipse.tensurastarlight.config.skills.aspectSeriesSkillConfig;
+import com.github.propheticeclipse.tensurastarlight.registry.skills.StarlightUniqueSkills;
+import com.hollingsworth.arsnouveau.setup.registry.ModEntities;
 import dev.shadowsoffire.apothic_attributes.api.ALObjects;
 import io.github.manasmods.manascore.config.ConfigRegistry;
 import io.github.manasmods.manascore.skill.api.ManasSkillInstance;
+import io.github.manasmods.tensura.ability.SkillUtils;
 import io.github.manasmods.tensura.ability.skill.Skill;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.stats.Stats;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeInstance;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.player.Player;
 
 public class wildenStalkerGeneticsSkill extends Skill {
 
@@ -29,6 +34,16 @@ public class wildenStalkerGeneticsSkill extends Skill {
 
     public int getAcquirementMastery(LivingEntity entity) {
         return CONFIG.acquirementMastery;
+    }
+
+    public boolean checkAcquiringRequirement(Player entity, double newEP) {
+
+        Boolean LightVestiges = SkillUtils.hasSkillFully(entity, StarlightUniqueSkills.LIGHT_REMAINS.get()) || SkillUtils.hasSkillFully(entity, StarlightUniqueSkills.VESTIGES_OF_EIDOLONS.get());
+        if (entity instanceof ServerPlayer servPlayer) {
+            return LightVestiges && (servPlayer.getStats().getValue(Stats.ENTITY_KILLED.get(ModEntities.WILDEN_STALKER.get())) >= 100);
+        } else {
+            return false;
+        }
     }
 
     public boolean canBeToggled(ManasSkillInstance instance, LivingEntity entity) {