+++ src/main/java/com/github/propheticeclipse/tensurastarlight/config/skills/malumSeriesSkillConfig.java
@@ -8,6 +8,7 @@ public class malumSeriesSkillConfig extends ManasConfig {
     public wardedSoul wardedSoul = new wardedSoul();
     public soulPlunderer soulPlunderer = new soulPlunderer();
     public malignantBody malignantBody = new malignantBody();
+    public spiritSorcerer spiritSorcerer = new spiritSorcerer();
 
     public String getFileName() {
         return "tensura/tensura-starlight/skills/malum-series";
@@ -56,5 +57,21 @@ public class malumSeriesSkillConfig extends ManasConfig {
 
         public double magicResistanceDamageModMastered = 0.75;
         public double magicResistanceDamageModUnmastered = 0.9;
+    }
+
+    public static class spiritSorcerer extends ManasSubConfig {
+        @Comment("How much MP should this skill cost to acquire/plunder/copy under normal circumstances?")
+        public double acquirementCost = 10000.0;
+        @Comment("How much mastery should this skill have by default?\nAny value above 1 = X Mastery at acquirement\n0 = Must be learned\nAny value below 0 = Requires X Learn Points")
+        public int acquirementMastery = 0;
+
+        public double soulWardMultiMastered = 0.5;
+        public double soulWardMultiUnmastered = 0.25;
+
+        public double soulWardIntegrityBonusMastered = 2.0;
+        public double soulWardIntegrityBonusUnmastered = 1.0;
+
+        public double soulAegisManaCost = 2500.0;
+        public int soulAegisCooldown = 5;
     }
 }