+++ src/main/java/com/github/propheticeclipse/tensurastarlight/registry/StarlightEffects.java
@@ -3,6 +3,7 @@ package com.github.propheticeclipse.tensurastarlight.registry;
 import com.github.propheticeclipse.tensurastarlight.effect.buff.EmberAttackImprovementEffect;
 import com.github.propheticeclipse.tensurastarlight.effect.buff.RemnantsOfAscensionArmorEffect;
 import com.github.propheticeclipse.tensurastarlight.effect.buff.ShapeForgottenArmorEffect;
+import com.github.propheticeclipse.tensurastarlight.effect.buff.SoulAegisEffect;
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
 import net.minecraft.core.registries.Registries;
@@ -13,15 +14,10 @@ public class StarlightEffects {
         MOB_EFFECTS.register();
     }
 
-    private static final DeferredRegister<MobEffect> MOB_EFFECTS;
-    public static final RegistrySupplier<MobEffect> EMBER_ATTACK_IMPROVEMENT;
-    public static final RegistrySupplier<MobEffect> SHAPE_FORGOTTEN_ARMOR;
-    public static final RegistrySupplier<MobEffect> REMNANT_ASCENSION_ARMOR;
+    private static final DeferredRegister<MobEffect> MOB_EFFECTS = DeferredRegister.create("trstarlight", Registries.MOB_EFFECT);
+    public static final RegistrySupplier<MobEffect> EMBER_ATTACK_IMPROVEMENT = MOB_EFFECTS.register("ember_attack_improvement", EmberAttackImprovementEffect::new);
+    public static final RegistrySupplier<MobEffect> SHAPE_FORGOTTEN_ARMOR = MOB_EFFECTS.register("shape_forgotten_armor", ShapeForgottenArmorEffect::new);
+    public static final RegistrySupplier<MobEffect> REMNANT_ASCENSION_ARMOR = MOB_EFFECTS.register("remnant_ascension_armor", RemnantsOfAscensionArmorEffect::new);
 
-    static {
-        MOB_EFFECTS = DeferredRegister.create("trstarlight", Registries.MOB_EFFECT);
-        EMBER_ATTACK_IMPROVEMENT = MOB_EFFECTS.register("ember_attack_improvement", EmberAttackImprovementEffect::new);
-        SHAPE_FORGOTTEN_ARMOR = MOB_EFFECTS.register("shape_forgotten_armor", ShapeForgottenArmorEffect::new);
-        REMNANT_ASCENSION_ARMOR = MOB_EFFECTS.register("remnant_ascension_armor", RemnantsOfAscensionArmorEffect::new);
-    }
+    public static final RegistrySupplier<MobEffect> SOUL_AEGIS = MOB_EFFECTS.register("soul_aegis", SoulAegisEffect::new);
 }