+++ src/main/java/com/github/propheticeclipse/tensurastarlight/registry/skills/StarlightExtraSkills.java
@@ -5,6 +5,7 @@ import com.github.propheticeclipse.tensurastarlight.ability.skill.extra.arsnouve
 import com.github.propheticeclipse.tensurastarlight.ability.skill.extra.arsnouveau.plasmaticSourceSkill;
 import com.github.propheticeclipse.tensurastarlight.ability.skill.extra.arsnouveau.sourcePoweredSkill;
 import com.github.propheticeclipse.tensurastarlight.ability.skill.extra.arsnouveau.wildenStalkSkill;
+import com.github.propheticeclipse.tensurastarlight.ability.skill.extra.malum.spiritSorcererSkill;
 import dev.architectury.registry.registries.RegistrySupplier;
 import io.github.manasmods.manascore.skill.api.ManasSkill;
 import io.github.manasmods.manascore.skill.impl.SkillRegistry;
@@ -13,12 +14,14 @@ import net.minecraft.resources.ResourceLocation;
 import java.util.function.Supplier;
 
 public class StarlightExtraSkills {
-    public static final RegistrySupplier<ManasSkill> DEAD_END_RAINBOW_SKILL = register("dead_end_rainbow_skill", deadEndRainbowSkill::new);
+    public static final RegistrySupplier<ManasSkill> DEAD_END_RAINBOW_SKILL = register("unleashed_dead_end", deadEndRainbowSkill::new);
     public static final RegistrySupplier<ManasSkill> ABNORMALLY_LUCKY = register("abnormally_lucky", abnormallyLuckySkill::new);
     public static final RegistrySupplier<ManasSkill> HYPER_METABOLISTIC = register("hyper_metabolistic", hyperMetabolisticSkill::new);
     public static final RegistrySupplier<ManasSkill> PLASMATIC_SOURCE = register("plasmatic_source", plasmaticSourceSkill::new);
     public static final RegistrySupplier<ManasSkill> SOURCE_POWERED = register("source_powered", sourcePoweredSkill::new);
     public static final RegistrySupplier<ManasSkill> WILDEN_STALK = register("wilden_stalk", wildenStalkSkill::new);
+
+    public static final RegistrySupplier<ManasSkill> SPIRIT_SORCERER = register("spirit_sorcerer", spiritSorcererSkill::new);
 
     private static <E extends ManasSkill> RegistrySupplier<E> register(String name, Supplier<E> supplier) {
         return SkillRegistry.SKILLS.register(ResourceLocation.fromNamespaceAndPath("trstarlight", name), supplier);